+++ src/components/landing/jumbotron/YOLOv7.tsx
@@ -163,7 +163,9 @@ export const YOLOv7 = () => {
 
   const handleGenrate = async () => {
     setSpinner(true);
-    swiper.autoplay.stop();
+    if (swiper) {
+      swiper.autoplay.stop();
+    }
     let imgString: string | null = "";
     // const defaultImage = await loadImageAndSetState("/images/yolo-default.png");
     if (input) {
@@ -339,7 +341,7 @@ export const YOLOv7 = () => {
         </div>
         <div className="mt-5 flex justify-end">
           <a
-            href="https://console.instill.tech/instill-wombat/pipelines/jumbotron-yolov7"
+            href="https://instill.tech/instill-wombat/pipelines/jumbotron-yolov7"
             target="_blank"
             className="absolute bottom-3 right-6 z-30 inline-flex items-center gap-x-2 divide-x divide-zinc-100/10 rounded bg-zinc-800/80 p-0 px-2 text-sm font-normal text-white no-underline drop-shadow-2xl backdrop-blur hover:text-blue-500 "
           >